novas componentizações nos scripts e melhorias na responsividade padrão do projeto
