+++ index.html
@@ -10,7 +10,6 @@
         crossorigin="anonymous" referrerpolicy="no-referrer" />
     <link rel="stylesheet" href="styles/index.css">
     <script type="module" src="scripts/index.js" defer></script>
-    <script src="scripts/dark-theme.js" defer></script>
 
     <title>Decodificador</title>
 </head>
@@ -51,7 +50,7 @@
                 </div>
             </section>
 
-            <section class="show__class__result" hidden>
+            <section class="show__class__result hide">
                 <p class="show__class__result__text"></p>
                 <button class="show__class__result__copy__button" type="button" id="copy-button">Copiar</button>
             </section>
