+++ styles/footer.css
@@ -1,12 +1,10 @@
 footer {
-    display: flex;
     background: var(--dark-blue-300);
     color: var(--white);
     font-size: 1.5rem;
     justify-content: center;
     align-items: center;
     text-align: center;
-    height: 14vh;
-    width: 100vw;
+    padding: 1.5rem;
     transition: 0.3s linear;
 }
