+++ styles/index.css
@@ -23,8 +23,6 @@
 }
 
 body {
-    width: 100%;
-    height: 100%;
     background: var(--light-blue-100);
     transition: 0.3s linear;
 }
@@ -127,10 +125,9 @@ body {
 
 main {
     display: flex;
-    width: 100vw;
-    height: auto;
+    width: 100%;
+    height: 100%;
     padding: 2%;
-    overflow: hidden;
 }
