+++ styles/input.css
@@ -2,14 +2,16 @@
     display: flex;
     margin: 4%;
     width: 100%;
-    height: 100vh;
+    height: 100%;
     flex-direction: column;
+    align-items: stretch;
 }
 
 .text__input__container__input {
-    display: flex;
     width: 100%;
-    height: 90vh;
+    height: 67vh;
+    resize: none;
+    box-sizing: border-box;
     border-radius: 24px;
     padding-top: 1rem;
     border: none;
